refactor: AnimatedLayout 컴포넌트화
